requirements.txt
facenet_pytorch
opencv-python-headless
pillow
boto3
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
